+++ docs/source/whats-new.rst
@@ -15,9 +15,143 @@ What's New
 
     np.random.seed(123456)
 
-.. _whats-new.v0.0.6:
+.. _whats-new.v0.0.12:
 
-v0.0.6 (unreleased)
+v0.0.12 (unreleased)
+--------------------
+
+New Features
+~~~~~~~~~~~~
+
+- Added a :py:func:`DataTree.level`, :py:func:`DataTree.depth`, and :py:func:`DataTree.width` property (:pull:`208`).
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+
+
+Breaking changes
+~~~~~~~~~~~~~~~~
+
+
+Deprecations
+~~~~~~~~~~~~
+
+Bug fixes
+~~~~~~~~~
+
+- Allow for altering of given dataset inside function called by :py:func:`map_over_subtree` (:issue:`188`, :pull:`194`).
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+
+Documentation
+~~~~~~~~~~~~~
+
+
+Internal Changes
+~~~~~~~~~~~~~~~~
+
+
+.. _whats-new.v0.0.11:
+
+v0.0.11 (01/09/2023)
+--------------------
+
+Big update with entirely new pages in the docs,
+new methods (``.drop_nodes``, ``.filter``, ``.leaves``, ``.descendants``), and bug fixes!
+
+New Features
+~~~~~~~~~~~~
+
+- Added a :py:meth:`DataTree.drop_nodes` method (:issue:`161`, :pull:`175`).
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+- New, more specific exception types for tree-related errors (:pull:`169`).
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+- Added a new :py:meth:`DataTree.descendants` property (:pull:`170`).
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+- Added a :py:meth:`DataTree.leaves` property (:pull:`177`).
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+- Added a :py:meth:`DataTree.filter` method (:pull:`184`).
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+
+Breaking changes
+~~~~~~~~~~~~~~~~
+
+- :py:meth:`DataTree.copy` copy method now only copies the subtree, not the parent nodes (:pull:`171`).
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+- Grafting a subtree onto another tree now leaves name of original subtree object unchanged (:issue:`116`, :pull:`172`, :pull:`178`).
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+- Changed the :py:meth:`DataTree.assign` method to just work on the local node (:pull:`181`).
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+
+Deprecations
+~~~~~~~~~~~~
+
+Bug fixes
+~~~~~~~~~
+
+- Fix bug with :py:meth:`DataTree.relative_to` method (:issue:`133`, :pull:`160`).
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+- Fix links to API docs in all documentation (:pull:`183`).
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+
+Documentation
+~~~~~~~~~~~~~
+
+- Changed docs theme to match xarray's main documentation. (:pull:`173`)
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+- Added ``Terminology`` page. (:pull:`174`)
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+- Added page on ``Working with Hierarchical Data`` (:pull:`179`)
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+- Added context content to ``Index`` page (:pull:`182`)
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+- Updated the README (:pull:`187`)
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+
+Internal Changes
+~~~~~~~~~~~~~~~~
+
+
+.. _whats-new.v0.0.10:
+
+v0.0.10 (12/07/2022)
+--------------------
+
+Adds accessors and a `.pipe()` method.
+
+New Features
+~~~~~~~~~~~~
+
+- Add the ability to register accessors on ``DataTree`` objects, by using ``register_datatree_accessor``. (:pull:`144`)
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+- Allow method chaining with a new :py:meth:`DataTree.pipe` method (:issue:`151`, :pull:`156`).
+  By `Justus Magin <https://github.com/keewis>`_.
+
+Breaking changes
+~~~~~~~~~~~~~~~~
+
+Deprecations
+~~~~~~~~~~~~
+
+Bug fixes
+~~~~~~~~~
+
+- Allow ``Datatree`` objects as values in :py:meth:`DataTree.from_dict` (:pull:`159`).
+  By `Justus Magin <https://github.com/keewis>`_.
+
+Documentation
+~~~~~~~~~~~~~
+
+- Added ``Reading and Writing Files`` page. (:pull:`158`)
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+
+Internal Changes
+~~~~~~~~~~~~~~~~
+
+- Avoid reading from same file twice with fsspec3 (:pull:`130`)
+  By `William Roberts <https://github.com/wroberts4>`_.
+
+
+.. _whats-new.v0.0.9:
+
+v0.0.9 (07/14/2022)
 -------------------
 
 New Features
@@ -34,9 +168,85 @@ Bug fixes
 
 Documentation
 ~~~~~~~~~~~~~
+- Switch docs theme (:pull:`123`).
+  By `JuliusBusecke <https://github.com/jbusecke>`_.
 
 Internal Changes
 ~~~~~~~~~~~~~~~~
+
+
+.. _whats-new.v0.0.7:
+
+v0.0.7 (07/11/2022)
+-------------------
+
+New Features
+~~~~~~~~~~~~
+
+- Improve the HTML repr by adding tree-style lines connecting groups and sub-groups (:pull:`109`).
+  By `Benjamin Woods <https://github.com/benjaminwoods>`_.
+
+Breaking changes
+~~~~~~~~~~~~~~~~
+
+- The ``DataTree.ds`` attribute now returns a view onto an immutable Dataset-like object, instead of an actual instance
+  of ``xarray.Dataset``. This make break existing ``isinstance`` checks or ``assert`` comparisons. (:pull:`99`)
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+
+Deprecations
+~~~~~~~~~~~~
+
+Bug fixes
+~~~~~~~~~
+
+- Modifying the contents of a ``DataTree`` object via the ``DataTree.ds`` attribute is now forbidden, which prevents
+  any possibility of the contents of a ``DataTree`` object and its ``.ds`` attribute diverging. (:issue:`38`, :pull:`99`)
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+- Fixed a bug so that names of children now always match keys under which parents store them (:pull:`99`).
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+
+Documentation
+~~~~~~~~~~~~~
+
+- Added ``Data Structures`` page describing the internal structure of a ``DataTree`` object, and its relation to
+  ``xarray.Dataset`` objects. (:pull:`103`)
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+- API page updated with all the methods that are copied from ``xarray.Dataset``. (:pull:`41`)
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+
+Internal Changes
+~~~~~~~~~~~~~~~~
+
+- Refactored ``DataTree`` class to store a set of ``xarray.Variable`` objects instead of a single ``xarray.Dataset``.
+  This approach means that the ``DataTree`` class now effectively copies and extends the internal structure of
+  ``xarray.Dataset``. (:pull:`41`)
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+- Refactored to use intermediate ``NamedNode`` class, separating implementation of methods requiring a ``name``
+  attribute from those not requiring it.
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+- Made ``testing.test_datatree.create_test_datatree`` into a pytest fixture (:pull:`107`).
+  By `Benjamin Woods <https://github.com/benjaminwoods>`_.
+
+
+
+.. _whats-new.v0.0.6:
+
+v0.0.6 (06/03/2022)
+-------------------
+
+Various small bug fixes, in preparation for more significant changes in the next version.
+
+Bug fixes
+~~~~~~~~~
+
+- Fixed bug with checking that assigning parent or new children did not create a loop in the tree (:pull:`105`)
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+- Do not call ``__exit__`` on Zarr store when opening (:pull:`90`)
+  By `Matt McCormick <https://github.com/thewtex>`_.
+- Fix netCDF encoding for compression (:pull:`95`)
+  By `Joe Hamman <https://github.com/jhamman>`_.
+- Added validity checking for node names (:pull:`106`)
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
 
 .. _whats-new.v0.0.5: